requirements.txt
#pip install -r requirements.txt
#python 3.9.12
#https://www.python.org/downloads/release/python-3912/
#https://phantomjs.org/download.html
selenium
beautifulsoup4
requests
python-telegram-bot